This code has also been updated to reflect the "No-UAC" version as it should
This has not been tested as yet, but this code should allow users with non-admin rights to install in areas of the system other than Root of C: Program Files, and Windows naturally. I am yet to confirm if the areas of the registry (being root) require admin rights? I have a feeling that they will, so this may require additional work?